Compilation warnings cleanup.
This commit is mostly comprised of unusued variables that were intentionally not used and/or forgotten during
development. One instance involved missing functionality (CountVoxels) but the new functionality is probably marginally
useful at best.

+++ src/Operations/ContourVote.cc
@@ -220,7 +220,7 @@ Drover ContourVote(const Drover& DICOM_data,
     //Create a new contour collection from the winning contours.
     contour_collection<double> cc_new;
     for(auto & cop : cop_ROIs){
-        if(cc_new.contours.size() >= WinnerCount) break;
+        if(static_cast<long int>(cc_new.contours.size()) >= WinnerCount) break;
         cc_new.contours.emplace_back(cop.get());
     }
